Conda install function (modified from reticulate)

R functions for managing Python conda environments.

conda_install(
  envname = NULL,
  packages,
  forge = FALSE,
  pip = FALSE,
  pip_ignore_installed = TRUE,
  conda = "auto",
  python_version = NULL,
  channel = NULL,
  ...
)

conda_python(envname = NULL, conda = "auto")

Arguments

envname
Name of conda environment

packages
Character vector with package names to install or remove.

forge
Include the Conda Forge repository.

pip
TRUE to use pip (defaults to FALSE)

pip_ignore_installed
Ignore installed versions when using pip. This is TRUE by default so that specific package versions can be installed even if they are downgrades. The FALSE option is useful for situations where you don't want a pip install to attempt an overwrite of a conda binary package (e.g. SciPy on Windows which is very difficult to install via pip due to compilation requirements).

conda
Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).

python_version
is the Python version to install. Python 3.4 is installed by default. But you could specify for instance: python_version="3.7".

channel
conda channel for pytorch

...
Optional arguments, reserved for future expansion.

Value

conda_list() returns a data frame with the names and paths to the respective python binaries of available environments. conda_create() returns the Path to the python binary of the created environment. conda_binary() returns the location of the main conda binary or NULL if none can be found.