Install PyTorch and its dependencies

Install PyTorch and its dependencies

install_pytorch(
  method = c("conda", "virtualenv", "auto"),
  conda = "auto",
  version = "default",
  envname = "r-torch",
  extra_packages = NULL,
  restart_session = TRUE,
  conda_python_version = "3.6",
  pip = FALSE,
  channel = "stable",
  cuda_version = NULL,
  dry_run = FALSE,
  ...
)

Arguments

method
Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows (as this isn't supported by PyTorch). Note also that since this command runs without privillege the "system" method is available only on Windows.

conda
The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda for more details.

version
PyTorch version to install. Specify "default" to install the CPU version of the latest release. Specify "gpu" to install the GPU version of the latest release.

You can also provide a full major.minor.patch specification (e.g. "1.1.0"), appending "-gpu" if you want the GPU version (e.g. "1.1.0-gpu").

Alternatively, you can provide the full URL to an installer binary (e.g. for a nightly binary).

envname
Name of Python environment to install within

extra_packages
Additional Python packages to install along with PyTorch.

restart_session
Restart R session after installing (note this will only occur within RStudio).

conda_python_version
the python version installed in the created conda environment. Python 3.6 is installed by default.

pip
logical

channel
conda channel

cuda_version
string for the cuda toolkit version

dry_run
logical, set to TRUE for unit tests, otherwise will run commands

...
other arguments passed to [reticulate::conda_install()] or [reticulate::virtualenv_install()].